requirements.txt
requests
beautifulsoup4
docopts
python-dotenv
ghapi
selenium
pandas
streamlit
GitPython
